chore(workflows): extract useDebouncedValue and LastSavedIndicator to lib
