+++ frontend/src/lib/components/LastSavedIndicator.tsx
@@ -1,0 +1,9 @@
+import { TZLabel } from 'lib/components/TZLabel'
+
+export function LastSavedIndicator({ timestamp }: { timestamp: string }): JSX.Element {
+    return (
+        <span className="text-xs text-tertiary">
+            Last saved <TZLabel time={timestamp} />
+        </span>
+    )
+}
