+++ products/workflows/frontend/Workflows/WorkflowScene.tsx
@@ -1,14 +1,14 @@
 import clsx from 'clsx'
 import { BindLogic, useActions, useValues } from 'kea'
 import { router } from 'kea-router'
-import { useEffect, useState } from 'react'
 
 import { IconInfo } from '@posthog/icons'
 import { LemonSwitch, Spinner, SpinnerOverlay } from '@posthog/lemon-ui'
 
 import { ActivityLog } from 'lib/components/ActivityLog/ActivityLog'
+import { LastSavedIndicator } from 'lib/components/LastSavedIndicator'
 import { NotFound } from 'lib/components/NotFound'
-import { TZLabel } from 'lib/components/TZLabel'
+import { useDebouncedValue } from 'lib/hooks/useDebouncedValue'
 import { LemonTab, LemonTabs } from 'lib/lemon-ui/LemonTabs'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { useAttachedLogic } from 'lib/logic/scenes/useAttachedLogic'
@@ -26,26 +26,6 @@ import { WorkflowLogs } from './WorkflowLogs'
 import { WorkflowMetrics } from './WorkflowMetrics'
 import { WorkflowSceneHeader } from './WorkflowSceneHeader'
 import { WorkflowSceneLogicProps, WorkflowTab, workflowSceneLogic } from './workflowSceneLogic'
-
-function useDebouncedValue<T>(value: T, delayMs: number): T {
-    const [debounced, setDebounced] = useState(value)
-    useEffect(() => {
-        if (value) {
-            const timer = setTimeout(() => setDebounced(value), delayMs)
-            return () => clearTimeout(timer)
-        }
-        setDebounced(value)
-    }, [value, delayMs])
-    return debounced
-}
-
-function LastSavedIndicator({ timestamp }: { timestamp: string }): JSX.Element {
-    return (
-        <span className="text-xs text-tertiary">
-            Last saved <TZLabel time={timestamp} />
-        </span>
-    )
-}
 
 export const scene: SceneExport<WorkflowSceneLogicProps> = {
     component: WorkflowScene,
